float32 to float 64 fix

+++ server/rpctypes/rpc_server_types.go
@@ -418,7 +418,7 @@ func (a Attributes) AreValid() error {
 	for _, val := range a {
 		if val.DisplayType != "" {
 			_, intOk := val.Value.(int)
-			_, floatOK := val.Value.(float32)
+			_, floatOK := val.Value.(float64)
 			if !intOk && !floatOK {
 				gamelog.L.Error().Err(fmt.Errorf("invalid attribute value %v for display type %s", val.Value, val.DisplayType)).Msg("invalid value in metadata")
 				errCount++
